Add Authentication (Auth.js)

+++ app/(api)/_datalib/_services/Users.ts
@@ -1,10 +1,13 @@
 import revalidateCache from '@actions/revalidateCache';
 import prisma from '../_prisma/client';
 import { UserInput } from '@datatypes/User';
+import { ApolloContext } from '@datalib/apolloServer';
 
 export default class Users {
   // CREATE
-  static async create(input: UserInput) {
+  static async create(input: UserInput, ctx: ApolloContext) {
+    if (!ctx.isOwner && !ctx.hasValidApiKey) return null;
+
     const user = await prisma.user.create({
       data: input,
     });
@@ -13,7 +16,9 @@ export default class Users {
   }
 
   // READ
-  static async find(id: string) {
+  static async find(id: string, ctx: ApolloContext) {
+    if (!ctx.isOwner && !ctx.hasValidApiKey) return null;
+
     return prisma.user.findUnique({
       where: {
         id,
@@ -21,7 +26,9 @@ export default class Users {
     });
   }
 
-  static async findMany(ids: string[]) {
+  static async findMany(ids: string[], ctx: ApolloContext) {
+    if (!ctx.isOwner && !ctx.hasValidApiKey) return null;
+
     if (!ids) {
       return prisma.user.findMany();
     }
@@ -36,7 +43,9 @@ export default class Users {
   }
 
   // UPDATE
-  static async update(id: string, input: UserInput) {
+  static async update(id: string, input: UserInput, ctx: ApolloContext) {
+    if (!ctx.isOwner && !ctx.hasValidApiKey) return null;
+
     try {
       const user = await prisma.user.update({
         where: {
@@ -52,7 +61,9 @@ export default class Users {
   }
 
   // DELETE
-  static async delete(id: string) {
+  static async delete(id: string, ctx: ApolloContext) {
+    if (!ctx.isOwner && !ctx.hasValidApiKey) return null;
+
     try {
       await prisma.user.delete({
         where: {
